Added our required variables for our WidgetController in a FStruct for better structure
Added a InitOverlay function which will add the Overlay widget to View[ort and Initialize the WidgetController required params, I called this function in HeroCharacter.cpp at void AHeroCharacter::InitAbilityActorInfo(). Because in this function we can get all the required params for initializing our Widget controller properly and this function will be called on client and server also.