Added a PreAttributeChange virtual void function in the Base_Attributeset class. It will be called before a attribute value is changed. It is useful for clamming a value So it will not go above or below a certain threshold
I will also test the PostGameplayEffectExecute() which will be called after changin a attribute value. It is more useful because sometimes there are some more calculations for attribute after calling the PreAttributeChange() function. So for being on the safe side we can also use the PostGameplayEffectExecute() virtual void overridden function.