Add CombatInterface with Virtual Function for PlayerLevel
Created a new CombatInterface that includes a virtual function responsible for returning the PlayerLevel variable. This variable aids in dynamically modifying attributes based on the player's level. Additionally, introduced Level variables in HeroPlayerState and EnemyCharacter. Implemented the CombatInterface in the parent class RPG_BaseCharacter and called the virtual function to retrieve and utilize the Player Level in both HeroCharacter and EnemyCharacter.