Added a Montage in our Ability which will trigger when we our ability is activated.
Added a Game Wait Gameplay Event which will trigger when we send a specific Gameplay Tag to it. It will be useful when we want to spawn our projectile in middle of a montage. We can trigger that Gameplay Event through a custom Anim notify whcih will take a input of Gameplay Tag and when it gets called through a montage it will send that Gameplay Event by which that Gameplay event will be executed.